feat(medication): Enhance stock status handling in medication components
