+++ packages/facility-server/app/routes/apiv1/medication.js
@@ -1675,6 +1675,7 @@ medication.get(
             'indication',
             'notes',
             'isVariableDose',
+            'date'
           ],
           required: true,
           include: [
