+++ packages/web/app/utils/medications.jsx
@@ -1,7 +1,16 @@
-import { ADMINISTRATION_FREQUENCY_SYNONYMS } from '@tamanu/constants';
+import React from 'react';
+import { Box } from '@mui/material';
+import {
+  ADMINISTRATION_FREQUENCY_SYNONYMS,
+  DRUG_STOCK_STATUS_LABELS,
+  DRUG_STOCK_STATUSES,
+} from '@tamanu/constants';
 import { camelCase } from 'lodash';
-import { formatTime } from '../components';
 import { getDateFromTimeString } from '@tamanu/shared/utils/medication';
+import { TableCellTag, ThemedTooltip, TranslatedText } from '@tamanu/ui-components';
+import styled from 'styled-components';
+import { formatTime, TranslatedEnum } from '../components';
+import { STOCK_STATUS_COLORS } from '../constants';
 
 export const getTranslatedFrequencySynonyms = (frequenciesEnabled, getTranslation) => {
   const result = {};
@@ -55,5 +64,48 @@ export const isWithinTimeSlot = (timeSlot, time, isFuture = false) => {
   }
 };
 
+const StyledTag = styled(TableCellTag)`
+  padding: 5px 12px;
+  border-radius: 999px;
+  font-size: 14px;
+  line-height: 18px;
+`;
+
 export const createPrescriptionHash = prescription =>
   `${prescription.medicationId}-${prescription.doseAmount}-${prescription.units}-${prescription.route}-${prescription.frequency}`;
+
+export const getStockStatus = ({ prescription }, useStyledTag = true) => {
+  const status =
+    prescription.medication?.referenceDrug?.facilities?.[0]?.stockStatus ||
+    DRUG_STOCK_STATUSES.UNKNOWN;
+  const quantity = prescription.medication?.referenceDrug?.facilities?.[0]?.quantity || 0;
+
+  const color = STOCK_STATUS_COLORS[status];
+
+  const content = useStyledTag ? (
+    <StyledTag $color={color} noWrap>
+      <TranslatedEnum value={status} enumValues={DRUG_STOCK_STATUS_LABELS} />
+    </StyledTag>
+  ) : (
+    <TranslatedEnum value={status} enumValues={DRUG_STOCK_STATUS_LABELS} />
+  );
+
+  if (status === DRUG_STOCK_STATUSES.IN_STOCK) {
+    return (
+      <ThemedTooltip
+        title={
+          <Box maxWidth="75px">
+            <TranslatedText
+              stringId="medication.stockLevel.tooltip"
+              fallback="Stock level: :quantity units"
+              replacements={{ quantity }}
+            />
+          </Box>
+        }
+      >
+        <span>{content}</span>
+      </ThemedTooltip>
+    );
+  }
+  return content;
+};
